% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attackEmail.R
\name{attackEmail}
\alias{attackEmail}
\title{Email Attacker}
\usage{
attackEmail(x, p = 0.8)
}
\arguments{
\item{x}{the email.}

\item{p}{probability of making attack.}
}
\value{
This function returns the attacked email.
}
\description{
Given a spam email, this function attacks it.
}
\examples{
attackEmail(x, p)
}
\keyword{atttacks}
\keyword{emails,}
