% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierUtilitiesGenerator.R
\name{classifierUtilitiesGenerator}
\alias{classifierUtilitiesGenerator}
\title{Classifier utilities generator}
\usage{
classifierUtilitiesGenerator(yc = 1, y = 1, ut = -10)
}
\arguments{
\item{yc}{defender's decision about email label, 1 spam 0 no spam.}

\item{y}{label of the email, 1 spam and 0 no spam.}
}
\description{
This function generates classifier utilities.
}
\examples{
classifierUtilitiesGenerator()
}
\keyword{emails,}
\keyword{random}
\keyword{utilities}
\keyword{utility,}
