% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costGenerator.R
\name{costGenerator}
\alias{costGenerator}
\title{Cost Generator}
\usage{
costGenerator(emailsDataset = trainingSetGenerator(1000))
}
\arguments{
\item{emailsDataset}{dataset containing the words from emails and whether the
emails are spam or not. The function will include the costs generated in 
emailsDataset.}
}
\description{
This function generates a costs associated to defender an atacker
}
\examples{
costGenerator()
}
\keyword{costs}
\keyword{emails,}
