% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltas.R
\name{deltas}
\alias{deltas}
\title{Computation of deltas.}
\usage{
deltas(r, var)
}
\arguments{
\item{r}{is the expected value of the beta distribution.}

\item{var}{is the variance of the beta distribution.}
}
\description{
This function computes the values of the deltas for the beta 
distribution.
}
\examples{
deltas(r, var)
}
\keyword{beta}
