% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getACRAProbabilities.R, R/main.R
\name{getACRAlabel}
\alias{getACRAlabel}
\alias{getACRAlabel}
\title{Get ACRA Probabilities}
\usage{
getACRAlabel(x, fit, var = 0.99, ut = matrix(c(1, -10, -1, 1), nrow = 2,
  byrow = T, dimnames = list(c("0", "1"), c("0", "1"))))

getACRAlabel(x, fit, var = 0.99, ut = matrix(c(1, -10, -1, 1), nrow = 2,
  byrow = T, dimnames = list(c("0", "1"), c("0", "1"))))
}
\arguments{
\item{x}{the email.}

\item{fit}{object of class naive-Bayes including the results from
training.}

\item{var}{Variance. Measure of certainty about attacker.}

\item{x}{the email.}

\item{fit}{object of class naive-Bayes including the results from
training.}

\item{var}{Variance. Measure of certainty about attacker.}
}
\value{
This function returns the ACRA probabilities.

This function returns the label of the possibly attacked email,
calculated with ACRA.
}
\description{
For a given possibly attacked email, and given 
parameters of NB classifier, this functions returns the ACRA probabilities,

For a given possibly attacked email, given utilities, and given 
parameters of NB classifier, this functions returns the label of the 
email, claculated using the ACRA machinery.
}
\examples{
getACRAprobabilities(x,fit,var)
getACRAlabel(x,fit,var)
}
\keyword{adversarial}
\keyword{attacks,}
\keyword{learning.}
