% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getACRAProbabilities.R
\name{getACRAprobabilities}
\alias{getACRAprobabilities}
\title{Get ACRA Probabilities}
\usage{
getACRAprobabilities(x, fit, var = 0.99)
}
\arguments{
\item{x}{the email.}

\item{fit}{object of class naive-Bayes including the results from
training.}

\item{var}{Variance. Measure of certainty about attacker.}
}
\value{
This function returns the ACRA probabilities.
}
\description{
For a given possibly attacked email, and given 
parameters of NB classifier, this functions returns the ACRA probabilities,
}
\examples{
getACRAprobabilities(x,fit,var)
}
\keyword{adversarial}
\keyword{attacks,}
\keyword{learning.}
