% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attackProbDistribution.R
\name{getAttackProbDsitribution}
\alias{getAttackProbDsitribution}
\title{Attacks Probability Distribution}
\usage{
getAttackProbDsitribution(x, y)
}
\arguments{
\item{x}{the email.}

\item{y}{the label of the email, 0 for innocent 1 for malicious.}
}
\value{
This function returns a vector with the probabilities of the 
possible attacks.
}
\description{
For a given email and a given label, this function computes the
probability distribution over the possible attacks of that email.
}
\examples{
getAttackProbDsitribution(x,y)
}
\keyword{attacks}
