% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBLabel.R
\name{getNBlabel}
\alias{getNBlabel}
\title{NB classificator with utilities}
\usage{
getNBlabel(x, fit, ut = matrix(c(1, -10, -1, 1), nrow = 2, byrow = T, dimnames
  = list(c("0", "1"), c("0", "1"))))
}
\arguments{
\item{x}{the email.}

\item{fit}{object of class naive-Bayes including the results from
training.}
}
\value{
This function returns the label of the email,
calculated with NB.
}
\description{
For a given email, given utilities, and given 
parameters of NB classifier, this functions returns the label of the 
email calculated using NB algorithm
}
\examples{
getNBlabel(x,fit,var)
}
\keyword{adversarial}
\keyword{attacks,}
\keyword{learning.}
