% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBPriors.R
\name{getNBPriors}
\alias{getNBPriors}
\title{Get Naive-Bayes Priors}
\usage{
getNBPriors(object)
}
\arguments{
\item{object}{an object of class naiveBayes containing at least the tables of the
likelihood of each word in the trained model.}
}
\value{
This function returns a matrix containg 
* \code{p(-)} = Under the label 0.
* \code{p(+)} = Under the label 1.
}
\description{
This function computes the Naive-Bayes estimates of the priors
\code{p(+)} and \code{p(-)}.
}
\keyword{naive-Bayes}
