% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNBProbabilities.R
\name{getNBProbabilities}
\alias{getNBProbabilities}
\title{Get Naive-Bayes Probabilities}
\usage{
getNBProbabilities(object, newdata, threshold = 0.001, eps = 0)
}
\arguments{
\item{object}{an object of class naiveBayes containing at least the tables of the
likelihood of each word in the trained model.}

\item{newdata}{a dataframe containing the x for which we wanto to calculate
\code{p(x|+)} and \code{p(x|-)}.}

\item{threshold}{value assigned to strictly probabilities <= eps.}

\item{eps}{Threshold for low probabilities.}
}
\value{
This function returns a matrix containg 
* \code{p(x|-)} = Under the label 0.
* \code{p(x|+)} = Under the label 1.
}
\description{
This function computes the Naive-Bayes estimates of
\code{p(x|+)} and \code{p(x|-)} for a given x.
}
\keyword{naive-Bayes}
