% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQs.R
\name{getQs}
\alias{getQs}
\title{Get q's of particular email}
\usage{
getQs(x, fit)
}
\arguments{
\item{x}{the email.}

\item{fit}{object of class naive-Bayes including the results from
training.}
}
\value{
This function returns a vector with the possible q's.
}
\description{
For a given email this function computes all the possible q's to use
later.
}
\examples{
getQs(x, fit)
}
\keyword{attacks}
