% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRa.R
\name{getRa}
\alias{getRa}
\title{Get ra from vector}
\usage{
getRa(emailMatrix, mail = c(1, 0, 1, 0, 1))
}
\arguments{
\item{emailMatrix}{dataframe that contains all the emails types}

\item{mail}{email to get ra.}
}
\value{
ra
}
\description{
Get ra from vector
}
\examples{
getRa(x)
}
\keyword{ra,}
