% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailGenerator.R
\name{mailGenerator}
\alias{mailGenerator}
\title{Email Generator}
\usage{
mailGenerator(k = 5, spam = T, subsetSpam = matrix(c(1, 0, 0, 0, 0, 1, 1,
  0, 0, 0), nrow = 2, byrow = T), probSubsetSpam = c(0.8, 0.2))
}
\arguments{
\item{k}{number of words in the email. Defaults is 5.}

\item{spam}{is the email spam? Defaults to TRUE.}

\item{subsetSpam}{matrix where each row is a configurations of words 
that have non-zero probability given that the email is spam. 
Defaults is \code{matrix(c(1,0,0,0,0,1,1,0,0,0),nrow = 2, byrow = T)}.}

\item{probSubsetSpam}{probability of the configurations in subsetSpam given
the email is spam. The probabilities for the rest of configurations are set
to 0. Defaults is \code{c(0.8,0.2)}.}
}
\value{
If \code{spam = T}, returns one of the rows of subsetSpam, according to 
the probabilities in probSubsetSopam. 
If \code{spam = F}, returns a k words email,
where each configuration is equally probable.
}
\description{
This function allows you to generate a k words email from a specific 
probability distribution, given the email is spam or not. 
Each word given the class of the email, is
generated from a Bernoulli distribution.
}
\examples{
mailGenerator()
}
\keyword{emails}
