% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromMailToNumber.R
\name{mailToNumber}
\alias{mailToNumber}
\title{Convert Mail to Number}
\usage{
mailToNumber(x)
}
\arguments{
\item{x}{the email}
}
\description{
For a given mail, this function gives the corresponding nunber.
}
\examples{
mailToNumber(x)
}
\keyword{mail}
\keyword{number}
\keyword{to}
