% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing to speed up computations}
\usage{
preprocessing(n, fit, words = c("viagra", "rajoy", "icmat", "hi", "bye"))
}
\arguments{
\item{n}{length of email.}

\item{fit}{object of class NB with the results of training.}

\item{words}{set of all words of the email.}
}
\value{
This function creates a dataframe including all possible
emails together with their means.
}
\description{
For a given email length, and set of words, this function generates
all possible emails, and the means to be used by ACRA.
}
\examples{
preprocessing(n, words, fit)
}
\keyword{preprocessing}
