% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing2v2.R
\name{preprocessing2V2}
\alias{preprocessing2V2}
\title{Get matrix of probabilities of classifier believing that the emails is labeled as malicious from all email types}
\usage{
preprocessing2V2(lengthEmail = 5, vectorWords = c("viagra", "rajoy",
  "icmat", "hi", "bye"), fit)
}
\arguments{
\item{lengthEmail}{legth of the email to consider.}

\item{vectorWords}{vector that contains the email words}

\item{fit}{object of class naive-Bayes including the results from
training.}
}
\value{
This function returns a matrix with all email types and their corresponding ra
}
\description{
Get matrix of probabilities of classifier believing that the emails is labeled as malicious from all email types
}
\examples{
preprocessing2V2(lengthEmail, vectorWords,fit)
}
\keyword{attacks,}
\keyword{ra,}
