% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randufGenerator.R
\name{randufGenerator}
\alias{randufGenerator}
\title{RANDUF generator}
\usage{
randufGenerator(y = 1, yc = 1)
}
\arguments{
\item{y}{label of the email, 1 spam and 0 no spam.}

\item{yc}{label of the email when defender make the classification of the email,
1 spam and 0 no spam}
}
\description{
This function generates random attacker utilities depending on expected attack 
result.
}
\examples{
randufGenerator()
}
\keyword{emails,}
\keyword{random}
\keyword{utilities}
\keyword{utility,}
