% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranprob.R
\name{ranprob}
\alias{ranprob}
\title{Get P_a(x) as seen by the adversary}
\usage{
ranprob(x, fit, var = 0.001)
}
\arguments{
\item{x}{the ATTACKED email.}

\item{fit}{object of class naive-Bayes including the results from
training.}
}
\value{
This function returns a P_a(x) as seen by the adversary
}
\description{
For a given attacked email this function computes the probability as
seen by the adversary that classifier classifies this email as spam
}
\examples{
ranprob(x, fit)
}
\keyword{attacks}
