% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingSetGenerator.R
\name{trainingSetGenerator}
\alias{trainingSetGenerator}
\title{Training Set Generator}
\usage{
trainingSetGenerator(n = 1000, spamPrev = 0.1, k = 5,
  Names = c("viagra", "rajoy", "icmat", "hi", "bye"),
  subsetSpam = matrix(c(1, 0, 0, 0, 0, 1, 1, 0, 0, 0), nrow = 2, byrow = T),
  probSubsetSpam = c(0.8, 0.2))
}
\arguments{
\item{n}{number of  emails to generate, defaults is 1000.}

\item{spamPrev}{spam prevalence, defaults is 0.1.}

\item{k}{number of words in each email. Defaults is 5.}

\item{Names}{vector with the names of the words in the email.}

\item{subsetSpam}{matrix where each row is a configurations of words 
that have non-zero probability given that the email is spam. 
Defaults is \code{matrix(c(1,0,0,0,0,1,1,0,0,0),nrow = 2, byrow = T)}.}

\item{probSubsetSpam}{probability of the configurations in subsetSpam given
the email is spam. The probabilities for the rest of configurations are set
to 0. Defaults is \code{c(0.8,0.2)}.}
}
\value{
This function returns a dataframe containing the generated emails.
}
\description{
This function generates a training set with n legal/spam emails, according
to the probability distributions in mailGenerator.
}
\examples{
trainingSetGenerator()
}
\keyword{emails}
