% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainRawNB.R
\name{trainRawNB}
\alias{trainRawNB}
\title{Train Raw Naive-Bayes}
\usage{
trainRawNB(trainingSet, words = c("viagra", "rajoy", "icmat", "hi", "bye"))
}
\arguments{
\item{trainingSet}{data frame containing all predictors and an extra column named
spam indicating wether each emai is legal or not. 1 == spam, 0 == nonspam.}

\item{words}{a string vector containing the words used as predictors.}
}
\value{
This function returns a list containg 
* apriori = The apriori distributions
* tables = Each word likelihoods.
}
\description{
This function trains a raw Naive Bayes in a given training set. In particular,
it calculates all relevant parameters such as likelihoods a aprior distributions.
}
\examples{
trainRawNB(trainingSet)
}
\keyword{naive-Bayes}
