% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getACRAlabel}
\alias{getACRAlabel}
\title{ACRA classificator}
\usage{
getACRAlabel(x, fit, var = 0.5, ut = -10)
}
\arguments{
\item{x}{the email.}

\item{fit}{object of class naive-Bayes including the results from
training.}

\item{var}{Variance. Measure of certainty about attacker.}
}
\value{
This function returns the label of the possibly attacked email,
calculated with ACRA.
}
\description{
For a given possibly attacked email, given utilities, and given 
parameters of NB classifier, this functions returns the label of the 
email, claculated using the ACRA machinery.
}
\examples{
getACRAlabel(x,fit,var)
}
\keyword{adversarial}
\keyword{attacks,}
\keyword{learning.}
