% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randut.R
\name{randut}
\alias{randut}
\title{Random Utility Generator}
\usage{
randut(y = 1, yc = 1)
}
\arguments{
\item{y}{label of the email, 1 spam and 0 no spam.}

\item{yc}{decision of defender about label of email,
1 spam and 0 no spam}
}
\description{
This function generates random attacker utilities depending on expected attack 
result.
}
\examples{
randut()
}
\keyword{emails,}
\keyword{random}
\keyword{utilities}
\keyword{utility,}
